Rename README files to clarify their contents.
This commit changes the names of two README files to better reflect the specific parts of the project they describe. The first file is renamed to README_DAY_1_PART_1.txt, and the second file is renamed to README_DAY_1_PART_2.txt. This should make it easier to navigate and understand the documentation structure.